refactor(documentForm): move components in appropriate folders
- small reusable components in /common
- DocumentForm in /appli
