+++ assets/js/react/components/appli/Document/DocumentForm/DocumentTypesHelper.js
