+++ assets/js/react/components/appli/Document/DocumentForm/formSteps/Step2.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { pathOr } from 'ramda';
 
-import Translate from '../../../Translate';
+import Translate from '../../../../common/Translate';
 
 import { DocumentFormContext } from '../Provider';
 import MassifAutoComplete from '../formElements/MassifAutoComplete';
